preprocess testing and model troubleshooting
The work book is a bit of a mess as I did a lot of work fixing the
preprocessing function to be less memory intensive. I also had to do a
lot of trouble shooting fine tuning the pretrained models in keras.
Apparently models with a batch normalization layers have problems.
specifically, the parameters in the batch normalization layers are not
frozen even if the layers are set to be non trainable. At the moment
there is no clear solution so features from the pretrained models will
need to be precomputed first and then fed to a top model.